[수정][#34] 홈 뷰 컴포넌트 분리
작업 완료
- totalView, inputFiledView 컴포넌트 분리 했습니다.
- transaction-history만 넣어서 render 하면 됩니다.
